refactor: make price in product table non-nullable and move ProductItemResponse in a different file
